+++ src/Modules/InventoryManagement/DigiDent.InventoryManagement.Domain/Requests/IRequestsCommandsRepository.cs
@@ -1,15 +1,10 @@
-﻿using DigiDent.InventoryManagement.Domain.Requests.ReadModels;
-using DigiDent.InventoryManagement.Domain.Requests.ValueObjects;
-using DigiDent.Shared.Kernel.ValueObjects.Pagination;
+﻿using DigiDent.InventoryManagement.Domain.Requests.ValueObjects;
 
 namespace DigiDent.InventoryManagement.Domain.Requests;
 
-public interface IRequestsRepository
+public interface IRequestsCommandsRepository
 {
     Task<Request?> GetByIdAsync(RequestId id, CancellationToken cancellationToken);
-    
-    Task<PaginatedResponse<RequestSummary>> GetAllAsync(
-        IPaginationOptions paginationOptions, CancellationToken cancellationToken);
     
     Task AddAsync(Request request, CancellationToken cancellationToken);
     Task UpdateAsync(Request request, CancellationToken cancellationToken);
